requirements.txt
pandas
matplotlib
seaborn
Pillow
requests
tqdm
numpy
torch
torchvision
timm
albumentations
opencv-python-headless